LPS-69093 Roll back API changes
