+++ modules/apps/web-experience/export-import/export-import-api/src/main/java/com/liferay/exportimport/content/processor/ExportImportContentProcessor.java
@@ -15,6 +15,7 @@
 package com.liferay.exportimport.content.processor;
 
 import com.liferay.exportimport.kernel.lar.PortletDataContext;
+import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.model.StagedModel;
 
 import java.io.Serializable;
@@ -23,18 +24,19 @@ import java.io.Serializable;
  * @author Gergely Mathe
  * @author Mate Thurzo
  */
-public interface ExportImportContentProcessor
-	<S extends StagedModel, T extends Serializable> {
+public interface ExportImportContentProcessor<T extends Serializable> {
 
 	public T replaceExportContentReferences(
-			PortletDataContext portletDataContext, S stagedModel, T content,
-			boolean exportReferencedContent, boolean escapeContent)
+			PortletDataContext portletDataContext, StagedModel stagedModel,
+			T content, boolean exportReferencedContent, boolean escapeContent)
 		throws Exception;
 
 	public T replaceImportContentReferences(
-			PortletDataContext portletDataContext, S stagedModel, T content)
+			PortletDataContext portletDataContext, StagedModel stagedModel,
+			T content)
 		throws Exception;
 
-	public boolean validateContentReferences(long groupId, T content);
+	public void validateContentReferences(long groupId, T content)
+		throws PortalException;
 
 }
