+++ modules/apps/web-experience/export-import/export-import-api/src/main/resources/com/liferay/exportimport/content/processor/packageinfo
@@ -1,1 +1,1 @@
-version 2.0.0
+version 1.0.0
