--- modules/apps/web-experience/export-import/export-import-service/src/main/java/com/liferay/exportimport/content/processor/base/BaseDocumentLibraryExportImportContentProcessor.java
@@ -1,563 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.exportimport.content.processor.base;
-
-import aQute.bnd.annotation.ProviderType;
-
-import com.liferay.document.library.kernel.exception.NoSuchFileEntryException;
-import com.liferay.document.library.kernel.model.DLFileEntry;
-import com.liferay.document.library.kernel.service.DLAppLocalServiceUtil;
-import com.liferay.document.library.kernel.service.DLFileEntryLocalServiceUtil;
-import com.liferay.document.library.kernel.util.DLUtil;
-import com.liferay.exportimport.content.processor.ExportImportContentProcessor;
-import com.liferay.exportimport.kernel.lar.ExportImportPathUtil;
-import com.liferay.exportimport.kernel.lar.PortletDataContext;
-import com.liferay.exportimport.kernel.lar.StagedModelDataHandlerUtil;
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.model.Company;
-import com.liferay.portal.kernel.model.Group;
-import com.liferay.portal.kernel.model.StagedModel;
-import com.liferay.portal.kernel.repository.model.FileEntry;
-import com.liferay.portal.kernel.service.CompanyLocalServiceUtil;
-import com.liferay.portal.kernel.service.GroupLocalServiceUtil;
-import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.ArrayUtil;
-import com.liferay.portal.kernel.util.CharPool;
-import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.HttpUtil;
-import com.liferay.portal.kernel.util.MapUtil;
-import com.liferay.portal.kernel.util.PortalUtil;
-import com.liferay.portal.kernel.util.PortletKeys;
-import com.liferay.portal.kernel.util.PredicateFilter;
-import com.liferay.portal.kernel.util.StringBundler;
-import com.liferay.portal.kernel.util.StringPool;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.kernel.xml.Element;
-
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-/**
- * @author Mate Thurzo
- */
-@ProviderType
-public abstract class BaseDocumentLibraryExportImportContentProcessor
-	<S extends StagedModel> implements ExportImportContentProcessor<S, String> {
-
-	@Override
-	public String replaceExportContentReferences(
-			PortletDataContext portletDataContext, S stagedModel,
-			String content, boolean exportReferencedContent,
-			boolean escapeContent)
-		throws Exception {
-
-		return replaceExportDLReferences(
-			portletDataContext, stagedModel, content, exportReferencedContent);
-	}
-
-	@Override
-	public String replaceImportContentReferences(
-			PortletDataContext portletDataContext, S stagedModel,
-			String content)
-		throws Exception {
-
-		return replaceImportDLReferences(
-			portletDataContext, stagedModel, content);
-	}
-
-	@Override
-	public boolean validateContentReferences(long groupId, String content) {
-		try {
-			validateDLReferences(groupId, content);
-
-			return true;
-		}
-		catch (Exception e) {
-			if (_log.isDebugEnabled()) {
-				_log.debug(e, e);
-			}
-
-			return false;
-		}
-	}
-
-	protected void deleteTimestampParameters(StringBuilder sb, int beginPos) {
-		beginPos = sb.indexOf(StringPool.CLOSE_BRACKET, beginPos);
-
-		if ((beginPos == -1) || (beginPos == (sb.length() - 1)) ||
-			(sb.charAt(beginPos + 1) != CharPool.QUESTION)) {
-
-			return;
-		}
-
-		int endPos = StringUtil.indexOfAny(
-			sb.toString(), DL_REFERENCE_LEGACY_STOP_CHARS, beginPos + 2);
-
-		if (endPos == -1) {
-			return;
-		}
-
-		String urlParams = sb.substring(beginPos + 1, endPos);
-
-		urlParams = HttpUtil.removeParameter(urlParams, "t");
-
-		sb.replace(beginPos + 1, endPos, urlParams);
-	}
-
-	protected Map<String, String[]> getDLReferenceParameters(
-		long groupId, String content, int beginPos, int endPos) {
-
-		boolean legacyURL = true;
-		char[] stopChars = DL_REFERENCE_LEGACY_STOP_CHARS;
-
-		if (content.startsWith("/documents/", beginPos)) {
-			legacyURL = false;
-			stopChars = DL_REFERENCE_STOP_CHARS;
-		}
-
-		endPos = StringUtil.indexOfAny(content, stopChars, beginPos, endPos);
-
-		if (endPos == -1) {
-			return null;
-		}
-
-		Map<String, String[]> map = new HashMap<>();
-
-		String dlReference = content.substring(beginPos, endPos);
-
-		while (dlReference.contains(StringPool.AMPERSAND_ENCODED)) {
-			dlReference = dlReference.replace(
-				StringPool.AMPERSAND_ENCODED, StringPool.AMPERSAND);
-		}
-
-		if (!legacyURL) {
-			String[] pathArray = dlReference.split(StringPool.SLASH);
-
-			if (pathArray.length < 3) {
-				return map;
-			}
-
-			map.put("groupId", new String[] {pathArray[2]});
-
-			if (pathArray.length == 4) {
-				map.put("uuid", new String[] {pathArray[3]});
-			}
-			else if (pathArray.length == 5) {
-				map.put("folderId", new String[] {pathArray[3]});
-				map.put(
-					"title", new String[] {HttpUtil.decodeURL(pathArray[4])});
-			}
-			else if (pathArray.length > 5) {
-				map.put("uuid", new String[] {pathArray[5]});
-			}
-		}
-		else {
-			dlReference = dlReference.substring(
-				dlReference.indexOf(CharPool.QUESTION) + 1);
-
-			map = HttpUtil.parameterMapFromString(dlReference);
-
-			String[] imageIds = null;
-
-			if (map.containsKey("img_id")) {
-				imageIds = map.get("img_id");
-			}
-			else if (map.containsKey("i_id")) {
-				imageIds = map.get("i_id");
-			}
-
-			imageIds = ArrayUtil.filter(
-				imageIds,
-				new PredicateFilter<String>() {
-
-					@Override
-					public boolean filter(String imageId) {
-						if (Validator.isNotNull(imageId)) {
-							return true;
-						}
-
-						return false;
-					}
-
-				});
-
-			if (ArrayUtil.isNotEmpty(imageIds)) {
-				map.put("image_id", imageIds);
-			}
-		}
-
-		map.put("endPos", new String[] {String.valueOf(endPos)});
-
-		String groupIdString = MapUtil.getString(map, "groupId");
-
-		if (groupIdString.equals("@group_id@")) {
-			groupIdString = String.valueOf(groupId);
-
-			map.put("groupId", new String[] {groupIdString});
-		}
-
-		return map;
-	}
-
-	protected FileEntry getFileEntry(Map<String, String[]> map) {
-		if (MapUtil.isEmpty(map)) {
-			return null;
-		}
-
-		FileEntry fileEntry = null;
-
-		try {
-			String uuid = MapUtil.getString(map, "uuid");
-			long groupId = MapUtil.getLong(map, "groupId");
-
-			if (Validator.isNotNull(uuid)) {
-				fileEntry = DLAppLocalServiceUtil.getFileEntryByUuidAndGroupId(
-					uuid, groupId);
-			}
-			else {
-				if (map.containsKey("folderId")) {
-					long folderId = MapUtil.getLong(map, "folderId");
-					String name = MapUtil.getString(map, "name");
-					String title = MapUtil.getString(map, "title");
-
-					if (Validator.isNotNull(title)) {
-						fileEntry = DLAppLocalServiceUtil.getFileEntry(
-							groupId, folderId, title);
-					}
-					else {
-						DLFileEntry dlFileEntry =
-							DLFileEntryLocalServiceUtil.fetchFileEntryByName(
-								groupId, folderId, name);
-
-						if (dlFileEntry != null) {
-							fileEntry = DLAppLocalServiceUtil.getFileEntry(
-								dlFileEntry.getFileEntryId());
-						}
-					}
-				}
-				else if (map.containsKey("image_id")) {
-					DLFileEntry dlFileEntry =
-						DLFileEntryLocalServiceUtil.fetchFileEntryByAnyImageId(
-							MapUtil.getLong(map, "image_id"));
-
-					if (dlFileEntry != null) {
-						fileEntry = DLAppLocalServiceUtil.getFileEntry(
-							dlFileEntry.getFileEntryId());
-					}
-				}
-			}
-		}
-		catch (Exception e) {
-			if (_log.isDebugEnabled()) {
-				_log.debug(e, e);
-			}
-			else if (_log.isWarnEnabled()) {
-				_log.warn(e.getMessage());
-			}
-		}
-
-		return fileEntry;
-	}
-
-	protected String replaceExportDLReferences(
-			PortletDataContext portletDataContext, S stagedModel,
-			String content, boolean exportReferencedContent)
-		throws Exception {
-
-		Group group = GroupLocalServiceUtil.getGroup(
-			portletDataContext.getGroupId());
-
-		if (group.isStagingGroup()) {
-			group = group.getLiveGroup();
-		}
-
-		if (group.isStaged() && !group.isStagedRemotely() &&
-			!group.isStagedPortlet(PortletKeys.DOCUMENT_LIBRARY)) {
-
-			return content;
-		}
-
-		StringBuilder sb = new StringBuilder(content);
-
-		String contextPath = PortalUtil.getPathContext();
-
-		String[] patterns = {
-			contextPath.concat("/c/document_library/get_file?"),
-			contextPath.concat("/documents/"),
-			contextPath.concat("/image/image_gallery?")
-		};
-
-		int beginPos = -1;
-		int endPos = content.length();
-
-		while (true) {
-			beginPos = StringUtil.lastIndexOfAny(content, patterns, endPos);
-
-			if (beginPos == -1) {
-				break;
-			}
-
-			Map<String, String[]> dlReferenceParameters =
-				getDLReferenceParameters(
-					portletDataContext.getScopeGroupId(), content,
-					beginPos + contextPath.length(), endPos);
-
-			FileEntry fileEntry = getFileEntry(dlReferenceParameters);
-
-			if (fileEntry == null) {
-				endPos = beginPos - 1;
-
-				continue;
-			}
-
-			endPos = MapUtil.getInteger(dlReferenceParameters, "endPos");
-
-			try {
-				if (exportReferencedContent) {
-					StagedModelDataHandlerUtil.exportReferenceStagedModel(
-						portletDataContext, stagedModel, fileEntry,
-						PortletDataContext.REFERENCE_TYPE_DEPENDENCY);
-				}
-				else {
-					Element entityElement =
-						portletDataContext.getExportDataElement(stagedModel);
-
-					portletDataContext.addReferenceElement(
-						stagedModel, entityElement, fileEntry,
-						PortletDataContext.REFERENCE_TYPE_DEPENDENCY, true);
-				}
-
-				String path = ExportImportPathUtil.getModelPath(fileEntry);
-
-				sb.replace(beginPos, endPos, "[$dl-reference=" + path + "$]");
-
-				deleteTimestampParameters(sb, beginPos);
-			}
-			catch (Exception e) {
-				if (_log.isDebugEnabled()) {
-					_log.debug(e, e);
-				}
-				else if (_log.isWarnEnabled()) {
-					StringBundler exceptionSB = new StringBundler(6);
-
-					exceptionSB.append("Unable to process file entry ");
-					exceptionSB.append(fileEntry.getFileEntryId());
-					exceptionSB.append(" for staged model ");
-					exceptionSB.append(stagedModel.getModelClassName());
-					exceptionSB.append(" with primary key ");
-					exceptionSB.append(stagedModel.getPrimaryKeyObj());
-
-					_log.warn(exceptionSB.toString());
-				}
-			}
-
-			endPos = beginPos - 1;
-		}
-
-		return sb.toString();
-	}
-
-	protected String replaceImportDLReferences(
-			PortletDataContext portletDataContext, S stagedModel,
-			String content)
-		throws Exception {
-
-		List<Element> referenceElements =
-			portletDataContext.getReferenceElements(
-				stagedModel, DLFileEntry.class);
-
-		for (Element referenceElement : referenceElements) {
-			long classPK = GetterUtil.getLong(
-				referenceElement.attributeValue("class-pk"));
-
-			Element referenceDataElement =
-				portletDataContext.getReferenceDataElement(
-					stagedModel, DLFileEntry.class, classPK);
-
-			String path = null;
-
-			if (referenceDataElement != null) {
-				path = referenceDataElement.attributeValue("path");
-			}
-
-			long groupId = GetterUtil.getLong(
-				referenceElement.attributeValue("group-id"));
-
-			if (Validator.isNull(path)) {
-				String className = referenceElement.attributeValue(
-					"class-name");
-
-				path = ExportImportPathUtil.getModelPath(
-					groupId, className, classPK);
-			}
-
-			if (!content.contains("[$dl-reference=" + path + "$]")) {
-				continue;
-			}
-
-			try {
-				StagedModelDataHandlerUtil.importReferenceStagedModel(
-					portletDataContext, stagedModel, DLFileEntry.class,
-					classPK);
-			}
-			catch (Exception e) {
-				if (_log.isDebugEnabled()) {
-					_log.debug(e, e);
-				}
-				else if (_log.isWarnEnabled()) {
-					StringBundler sb = new StringBundler(6);
-
-					sb.append("Unable to process file entry ");
-					sb.append(classPK);
-					sb.append(" for ");
-					sb.append(stagedModel.getModelClassName());
-					sb.append(" with primary key ");
-					sb.append(stagedModel.getPrimaryKeyObj());
-
-					_log.warn(sb.toString());
-				}
-			}
-
-			Map<Long, Long> dlFileEntryIds =
-				(Map<Long, Long>)portletDataContext.getNewPrimaryKeysMap(
-					DLFileEntry.class);
-
-			long fileEntryId = MapUtil.getLong(
-				dlFileEntryIds, classPK, classPK);
-
-			FileEntry importedFileEntry = null;
-
-			try {
-				importedFileEntry = DLAppLocalServiceUtil.getFileEntry(
-					fileEntryId);
-			}
-			catch (PortalException pe) {
-				if (_log.isDebugEnabled()) {
-					_log.debug(pe, pe);
-				}
-				else if (_log.isWarnEnabled()) {
-					_log.warn(pe.getMessage());
-				}
-
-				continue;
-			}
-
-			String url = DLUtil.getPreviewURL(
-				importedFileEntry, importedFileEntry.getFileVersion(), null,
-				StringPool.BLANK, false, false);
-
-			if (url.contains(StringPool.QUESTION)) {
-				content = StringUtil.replace(content, "$]?", "$]&");
-			}
-
-			content = StringUtil.replace(
-				content, "[$dl-reference=" + path + "$]", url);
-		}
-
-		return content;
-	}
-
-	protected void validateDLReferences(long groupId, String content)
-		throws PortalException {
-
-		String portalURL = PortalUtil.getPathContext();
-
-		ServiceContext serviceContext =
-			ServiceContextThreadLocal.getServiceContext();
-
-		if ((serviceContext != null) &&
-			(serviceContext.getThemeDisplay() != null)) {
-
-			ThemeDisplay themeDisplay = serviceContext.getThemeDisplay();
-
-			portalURL =
-				PortalUtil.getPortalURL(themeDisplay) +
-					PortalUtil.getPathContext();
-		}
-
-		String[] patterns = {
-			portalURL.concat("/c/document_library/get_file?"),
-			portalURL.concat("/documents/"),
-			portalURL.concat("/image/image_gallery?")
-		};
-
-		long[] companyIds = PortalUtil.getCompanyIds();
-
-		String[] completePatterns =
-			new String[patterns.length * companyIds.length];
-
-		int i = 0;
-
-		for (long companyId : companyIds) {
-			Company company = CompanyLocalServiceUtil.getCompany(companyId);
-
-			String webId = company.getWebId();
-
-			for (String pattern : patterns) {
-				completePatterns[i] = webId.concat(pattern);
-
-				i++;
-			}
-		}
-
-		int beginPos = -1;
-		int endPos = content.length();
-
-		while (true) {
-			beginPos = StringUtil.lastIndexOfAny(
-				content, completePatterns, endPos);
-
-			if (beginPos == -1) {
-				break;
-			}
-
-			Map<String, String[]> dlReferenceParameters =
-				getDLReferenceParameters(
-					groupId, content, beginPos + portalURL.length(), endPos);
-
-			FileEntry fileEntry = getFileEntry(dlReferenceParameters);
-
-			if (fileEntry == null) {
-				throw new NoSuchFileEntryException(
-					"File entry reference parameters: " +
-						MapUtil.toString(dlReferenceParameters));
-			}
-
-			endPos = beginPos - 1;
-		}
-	}
-
-	protected static final char[] DL_REFERENCE_LEGACY_STOP_CHARS = {
-		CharPool.APOSTROPHE, CharPool.CLOSE_BRACKET, CharPool.CLOSE_CURLY_BRACE,
-		CharPool.CLOSE_PARENTHESIS, CharPool.GREATER_THAN, CharPool.LESS_THAN,
-		CharPool.PIPE, CharPool.QUOTE, CharPool.SPACE
-	};
-
-	protected static final char[] DL_REFERENCE_STOP_CHARS = {
-		CharPool.APOSTROPHE, CharPool.CLOSE_BRACKET, CharPool.CLOSE_CURLY_BRACE,
-		CharPool.CLOSE_PARENTHESIS, CharPool.GREATER_THAN, CharPool.LESS_THAN,
-		CharPool.PIPE, CharPool.QUESTION, CharPool.QUOTE, CharPool.SPACE
-	};
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		BaseDocumentLibraryExportImportContentProcessor.class);
-
-}
